incorporating review comments
