+++ code/utils.py
@@ -38,12 +38,10 @@ def populate_userid_mapping():
         user_id_lookup[m[0]] = int(m[1])
 
 
-def plot_with_tsne(email_ids, embeddings, display_plot):
+def plot_with_tsne(labels, embeddings, display_plot=True):
     """
     expects a list of email_ids and numpy ndarray of embeddings. The numpy ndarray should have shape L,D where D is the
     size of embeddings and L is the number of users
-    :param email2embedding:
-    :return:
     """
     tsne = TSNE(verbose=1, method='exact')
     start = time.time()
@@ -54,7 +52,7 @@ def plot_with_tsne(email_ids, embeddings, display_plot):
     if display_plot:
         fig, ax = plt.subplots()
         scatter = ax.scatter(tsne_embs[:,0], tsne_embs[:,1], s=30)
-        tooltip = mpld3.plugins.PointLabelTooltip(scatter, labels=email_ids)
+        tooltip = mpld3.plugins.PointLabelTooltip(scatter, labels=labels)
         mpld3.plugins.connect(fig, tooltip)
         mpld3.show()
